requirements.txt
open3d==0.18.0
pillow==10.2.0
mediapy==1.1.2
lpips==0.1.4
scikit-image==0.21.0
tqdm==4.66.2
trimesh==4.3.2
plyfile
opencv-python
huggingface_hub
diffusers
accelerate
transformers
ConfigArgParse
tensorboard
timm
matplotlib==3.8.0
natsort
pytorch-fid
torchmetrics
bitsandbytes

einops==0.6.1
pytorch_lightning==1.5.0
open_clip_torch==2.16.0
pie-torch
